DNS-Agent-Agno/pyproject.toml
[project]
name = "dns-agent-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno[aws]>=1.5.1",
    "dnspython>=2.7.0",
    "duckduckgo-search>=8.0.2",
    "python-dotenv>=1.1.0",
    "python-whois>=0.9.5",
    "streamlit>=1.45.1",
]
